+++ platforms/ios/www/js/index.js
@@ -35,6 +35,7 @@ var app = {
     onDeviceReady: function() {
         app.receivedEvent('deviceready');
         app.pluginInitialize();
+        socket.init();
     },
     // Update DOM on a Received Event
     receivedEvent: function(id) {
@@ -113,6 +114,7 @@ var app = {
             counter++;
 
             console.log('Running since ' + counter + ' sec');
+            socket.doSend('Running since ' + counter + ' sec');
 
             cordova.plugins.notification.badge.set(counter);
 
@@ -131,6 +133,46 @@ var app = {
     onModeDeactivated: function() {
         cordova.plugins.notification.badge.clear();
         clearInterval(app.timer);
+    }
+};
+
+var socket = {
+    _socket: null,
+
+    init: function() {
+        if (!window.hasOwnProperty('WebSocket'))
+            return;
+
+        this._socket = new WebSocket('ws://echo.websocket.org/');
+
+        this._socket.onopen    = function(evt) { socket.onOpen(evt); };
+        this._socket.onclose   = function(evt) { socket.onClose(evt); };
+        this._socket.onmessage = function(evt) { socket.onMessage(evt); };
+        this._socket.onerror   = function(evt) { socket.onError(evt); };
+    },
+
+    onOpen: function(evt) {
+        console.log('CONNECTED');
+        this.doSend('background-mode plugin rocks');
+    },
+
+    onClose: function(evt) {
+        console.log('DISCONNECTED');
+    },
+
+    onMessage: function(evt) {
+        console.log('RECEIVED: ' + evt.data);
+    },
+
+    onError: function(evt) {
+        console.log('ERROR: ' + evt.data);
+    },
+
+    doSend: function(message) {
+        if (this._socket) {
+            console.log('SENT: ' + message);
+            this._socket.send(message);
+        }
     }
 };