+++ platforms/windows/www/cordova_plugins.js
@@ -1,36 +1,5 @@
-﻿cordova.define('cordova/plugin_list', function(require, exports, module) {
+cordova.define('cordova/plugin_list', function(require, exports, module) {
 module.exports = [
-    {
-        "id": "cordova-plugin-device.device",
-        "file": "plugins/cordova-plugin-device/www/device.js",
-        "pluginId": "cordova-plugin-device",
-        "clobbers": [
-            "device"
-        ]
-    },
-    {
-        "id": "cordova-plugin-device.DeviceProxy",
-        "file": "plugins/cordova-plugin-device/src/windows/DeviceProxy.js",
-        "pluginId": "cordova-plugin-device",
-        "merges": [
-            ""
-        ]
-    },
-    {
-        "id": "cordova-plugin-background-mode.BackgroundMode",
-        "file": "plugins/cordova-plugin-background-mode/www/background-mode.js",
-        "pluginId": "cordova-plugin-background-mode",
-        "clobbers": [
-            "cordova.plugins.backgroundMode",
-            "plugin.backgroundMode"
-        ]
-    },
-    {
-        "id": "cordova-plugin-background-mode.BackgroundMode.Proxy",
-        "file": "plugins/cordova-plugin-background-mode/src/windows/BackgroundModeProxy.js",
-        "pluginId": "cordova-plugin-background-mode",
-        "runs": true
-    },
     {
         "id": "cordova-plugin-badge.Badge",
         "file": "plugins/cordova-plugin-badge/www/badge.js",
@@ -76,17 +45,42 @@ module.exports = [
         "file": "plugins/cordova-plugin-vibration/src/windows/VibrationProxy.js",
         "pluginId": "cordova-plugin-vibration",
         "runs": true
+    },
+    {
+        "id": "cordova-plugin-device.device",
+        "file": "plugins/cordova-plugin-device/www/device.js",
+        "pluginId": "cordova-plugin-device",
+        "clobbers": [
+            "device"
+        ]
+    },
+    {
+        "id": "cordova-plugin-device.DeviceProxy",
+        "file": "plugins/cordova-plugin-device/src/windows/DeviceProxy.js",
+        "pluginId": "cordova-plugin-device",
+        "merges": [
+            ""
+        ]
+    },
+    {
+        "id": "cordova-plugin-background-mode.BackgroundMode",
+        "file": "plugins/cordova-plugin-background-mode/www/background-mode.js",
+        "pluginId": "cordova-plugin-background-mode",
+        "clobbers": [
+            "cordova.plugins.backgroundMode",
+            "plugin.backgroundMode"
+        ]
     }
 ];
 module.exports.metadata = 
 // TOP OF METADATA
 {
     "cordova-plugin-app-event": "1.2.0",
-    "cordova-plugin-device": "1.1.4",
-    "cordova-plugin-background-mode": "0.7.0",
     "cordova-plugin-badge": "0.7.4",
     "cordova-plugin-console": "1.0.5",
-    "cordova-plugin-vibration": "2.1.3"
+    "cordova-plugin-vibration": "2.1.3",
+    "cordova-plugin-device": "1.1.4",
+    "cordova-plugin-background-mode": "0.7.2"
 };
 // BOTTOM OF METADATA
 });