+++ platforms/windows/www/plugins/cordova-plugin-badge/www/badge.js
@@ -1,4 +1,4 @@
-﻿cordova.define("cordova-plugin-badge.Badge", function(require, exports, module) {
+cordova.define("cordova-plugin-badge.Badge", function(require, exports, module) {
 /*
  * Copyright (c) 2013-2016 by appPlant GmbH. All rights reserved.
  *